@@ -18,8 +18,10 @@
     "prebuild": "npm run clean",
     "build": "tsc --project tsconfig.build.json  && npm run assets:copy",
     "start": "npm run build && cd dist && node ./index.js",
+    "start:dev": "npm run build && cd dist && node --enable-source-maps ./index.js",
     "assets:copy": "copyfiles -f ./config/* ./dist/config && copyfiles -f ./openapi3.yaml ./dist/ && copyfiles ./package.json dist",
-    "clean": "rimraf dist"
+    "clean": "rimraf dist",
+    "install": "npx husky install"
   },
   "directories": {
     "test": "tests"
@@ -38,51 +40,55 @@
     }
   },
   "dependencies": {
-    "@godaddy/terminus": "^4.10.2",
+    "@godaddy/terminus": "^4.12.1",
     "@map-colonies/error-express-handler": "^2.1.0",
-    "@map-colonies/express-access-log-middleware": "^1.0.0",
-    "@map-colonies/js-logger": "0.0.5",
-    "@map-colonies/openapi-express-viewer": "^2.0.1",
-    "@map-colonies/telemetry": "4.1.0",
-    "@opentelemetry/api": "^1.1.0",
-    "@opentelemetry/api-metrics": "^0.29.0",
+    "@map-colonies/express-access-log-middleware": "^2.0.1",
+    "@map-colonies/js-logger": "^1.0.1",
+    "@map-colonies/openapi-express-viewer": "^3.0.0",
+    "@map-colonies/read-pkg": "0.0.1",
+    "@map-colonies/telemetry": "^5.0.0",
+    "@opentelemetry/api": "^1.4.1",
+    "@opentelemetry/api-metrics": "0.23.0",
+    "@opentelemetry/instrumentation-express": "0.32.1",
+    "@opentelemetry/instrumentation-http": "0.35.1",
     "compression": "^1.7.4",
-    "config": "^3.3.7",
-    "express": "^4.18.1",
-    "express-openapi-validator": "^4.13.8",
+    "config": "^3.3.9",
+    "express": "^4.18.2",
+    "express-openapi-validator": "^5.0.4",
     "http-status-codes": "^2.2.0",
     "reflect-metadata": "^0.1.13",
-    "tsyringe": "^4.7.0"
+    "tsyringe": "^4.8.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.0.2",
-    "@commitlint/config-conventional": "^17.0.2",
-    "@map-colonies/eslint-config": "^3.1.1",
+    "@commitlint/cli": "^17.6.6",
+    "@commitlint/config-conventional": "^17.6.6",
+    "@map-colonies/eslint-config": "^3.1.3",
     "@map-colonies/prettier-config": "0.0.1",
     "@map-colonies/standard-version-update-helm-version": "^2.0.1",
     "@map-colonies/standard-version-update-openapi3-version": "0.0.2",
     "@redocly/openapi-cli": "^1.0.0-beta.94",
     "@types/compression": "^1.7.2",
-    "@types/config": "0.0.41",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^27.5.1",
+    "@types/config": "^3.3.0",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.2",
     "@types/multer": "^1.4.7",
     "@types/supertest": "^2.0.12",
-    "@types/yamljs": "^0.2.31",
-    "commitlint": "^17.0.2",
+    "@types/swagger-ui-express": "^4.1.3",
+    "commitlint": "^17.6.6",
     "copyfiles": "^2.4.1",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.16.0",
-    "husky": "^8.0.1",
-    "jest": "^28.1.0",
+    "eslint": "^8.43.0",
+    "husky": "^8.0.3",
+    "jest": "^29.5.0",
     "jest-create-mock-instance": "^2.0.0",
-    "jest-html-reporters": "^3.0.8",
-    "prettier": "^2.6.2",
+    "jest-html-reporters": "^3.1.4",
+    "jest-openapi": "^0.14.2",
+    "prettier": "^2.8.8",
     "pretty-quick": "^3.1.3",
-    "rimraf": "^3.0.2",
+    "rimraf": "^5.0.1",
     "standard-version": "^9.5.0",
-    "supertest": "^6.2.3",
-    "ts-jest": "^28.0.3",
-    "typescript": "^4.3.4"
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.5",
+    "typescript": "^5.1.3"
   }
 }
