@@ -1,5 +1,5 @@
 /* eslint-disable import/first */
-// this import must be called before the first import of tsyring
+// this import must be called before the first import of tsyringe
 import 'reflect-metadata';
 import './common/tracing';
 import { createServer } from 'http';
@@ -7,20 +7,15 @@ import { createTerminus } from '@godaddy/terminus';
 import { Logger } from '@map-colonies/js-logger';
 import { container } from 'tsyringe';
 import config from 'config';
-import { DEFAULT_SERVER_PORT, Services } from './common/constants';
+import { DEFAULT_SERVER_PORT, SERVICES } from './common/constants';
 
 import { getApp } from './app';
 
-interface IServerConfig {
-  port: string;
-}
-
-const serverConfig = config.get<IServerConfig>('server');
-const port: number = parseInt(serverConfig.port) || DEFAULT_SERVER_PORT;
+const port: number = config.get<number>('server.port') || DEFAULT_SERVER_PORT;
 
 const app = getApp();
 
-const logger = container.resolve<Logger>(Services.LOGGER);
+const logger = container.resolve<Logger>(SERVICES.LOGGER);
 const stubHealthcheck = async (): Promise<void> => Promise.resolve();
 // eslint-disable-next-line @typescript-eslint/naming-convention
 const server = createTerminus(createServer(app), { healthChecks: { '/liveness': stubHealthcheck }, onSignal: container.resolve('onSignal') });
